+++ frontend/src/hooks/queries/useFetchThreads.ts
@@ -11,11 +11,11 @@ export const useFetchThreads = (teamPlaceId: number) => {
     ['threadData', teamPlaceId],
     ({ pageParam = undefined }) => fetchThreads(teamPlaceId, pageParam),
     {
+      enabled: teamPlaceId > 0,
       getNextPageParam: (lastPage) => {
         if (lastPage.threads.length !== THREAD_SIZE) return undefined;
         return lastPage.threads[THREAD_SIZE - 1].id;
       },
-      enabled: teamPlaceId > 0,
     },
   );
